requirements_new.txt
kafka-python
google-cloud-storage
pandas
requests
protobuf
PyYAML
tenacity
venv-pack 
kafka-python
confluent-kafka